@@ -7,8 +7,8 @@
   "private": true,
   "engineStrict": true,
   "engines": {
-    "node": ">= 16",
-    "npm": ">= 8"
+    "node": ">= 20",
+    "npm": ">= 9"
   },
   "repository": {
     "type": "git",
@@ -19,40 +19,30 @@
     "url": "https://www.saasmakers.club/"
   },
   "scripts": {
-    "dev": "nuxt",
+    "dev": "nuxt dev",
     "build": "nuxt generate",
-    "clean": "rm -rf ./dist",
+    "clean": "rm -rf ./build",
     "test": "npm run lint",
-    "lint:js": "eslint --ext \".js,.vue\" --ignore-path .gitignore .",
-    "lint:prettier": "prettier --check .",
+    "lint:js": "eslint .",
+    "lint:prettier": "prettier --check src/",
     "lint": "npm run lint:js && npm run lint:prettier"
   },
   "dependencies": {
-    "core-js": "3.19.3",
-    "nuxt": "2.17.1",
-    "vue": "2.7.14",
-    "vue-server-renderer": "2.7.14",
-    "vue-template-compiler": "2.7.14",
-    "vue-observe-visibility": "1.0.0",
-    "webpack": "4.46.0",
+    "vue": "3.4.31",
+    "vue3-observe-visibility": "1.0.1",
     "keycode": "2.2.1"
   },
   "devDependencies": {
-    "@babel/eslint-parser": "7.16.3",
-    "@nuxtjs/eslint-config": "8.0.0",
-    "@nuxtjs/eslint-module": "3.0.2",
-    "@nuxtjs/svg": "0.4.0",
-    "@nuxtjs/robots": "2.5.0",
-    "@nuxtjs/sitemap": "2.4.0",
-    "@nuxtjs/style-resources": "1.2.1",
-    "pug": "3.0.2",
-    "sass": "1.49.11",
+    "@nuxt/eslint-config": "0.3.13",
+    "@nuxtjs/robots": "3.0.0",
+    "@nuxtjs/sitemap": "5.2.2",
+    "pug": "3.0.3",
+    "sass": "1.77.6",
     "pug-plain-loader": "1.1.0",
-    "sass-loader": "10.2.1",
-    "eslint": "8.4.1",
-    "eslint-config-prettier": "8.3.0",
-    "eslint-plugin-nuxt": "3.1.0",
-    "eslint-plugin-vue": "8.2.0",
+    "sass-loader": "14.2.1",
+    "eslint": "9.6.0",
+    "nuxt": "3.12.2",
+    "nuxt-svgo": "4.0.1",
     "prettier": "2.5.1",
     "lodash.merge": "4.6.2"
   }
