@@ -14,6 +14,7 @@ import fs from "fs";
 import path from "path";
 
 // NPM
+import { defineNuxtConfig } from "nuxt/config";
 import merge from "lodash.merge";
 
 // CONFIGURATION
@@ -52,222 +53,194 @@ const CONFIG = (function () {
  * EXPORTS
  * ************************************************************************* */
 
-export default {
-  // Target: https://go.nuxtjs.dev/config-target
+export default defineNuxtConfig({
+  // Server-Side Rendering (SSG)
+  ssr: true,
   target: "static",
 
-  // Global name: \
-  //   https://nuxtjs.org/docs/configuration-glossary/configuration-global-name
-  globalName: "saasmakers",
-
-  // Server-Side Rendering: \
-  //   https://nuxtjs.org/docs/configuration-glossary/configuration-ssr
-  ssr: true,
-
-  // Telemetry: \
-  //   https://nuxtjs.org/docs/configuration-glossary/configuration-telemetry
+  // Telemetry
   telemetry: false,
 
-  // Global page headers: https://go.nuxtjs.dev/config-head
-  head: {
-    title: "saasmakers-web",
-
-    htmlAttrs: {
-      lang: "en"
-    },
-
-    meta: [
-      {
-        charset: "utf-8"
-      },
-
-      {
-        name: "viewport",
-        content: "width=device-width, initial-scale=1"
-      },
-
-      {
-        name: "format-detection",
-        content: "telephone=no"
-      },
-
-      {
-        hid: "description",
-        name: "description",
-        content:
-          "Chaque mois, nous parlons d’un sujet SaaS avec un expert du domaine."
-      }
-    ],
-
-    link: [
-      {
-        rel: "shortcut icon",
-        type: "image/x-icon",
-        href: "/favicons/favicon.ico"
-      },
-
-      {
-        rel: "icon",
-        type: "image/png",
-        href: "/favicons/favicon.png"
-      },
-
-      {
-        rel: "icon",
-        type: "image/png",
-        sizes: "512x512",
-        href: "/favicons/favicon-512x512.png"
-      },
-
-      {
-        rel: "icon",
-        type: "image/png",
-        sizes: "256x256",
-        href: "/favicons/favicon-256x256.png"
-      },
-
-      {
-        rel: "icon",
-        type: "image/png",
-        sizes: "192x192",
-        href: "/favicons/favicon-192x192.png"
-      },
-
-      {
-        rel: "icon",
-        type: "image/png",
-        sizes: "144x144",
-        href: "/favicons/favicon-144x144.png"
-      },
-
-      {
-        rel: "icon",
-        type: "image/png",
-        sizes: "128x128",
-        href: "/favicons/favicon-128x128.png"
-      },
-
-      {
-        rel: "icon",
-        type: "image/png",
-        sizes: "32x32",
-        href: "/favicons/favicon-32x32.png"
-      }
-    ]
+  // General Site Configuration
+  site: {
+    url: CONFIG.url.saasmakers_web
   },
 
-  // Source directory: \
-  //   https://nuxtjs.org/docs/configuration-glossary/configuration-srcdir
+  // Source Directory
   srcDir: "src/",
+  publicDir: "public/",
 
-  // Directories: \
-  //   https://nuxtjs.org/docs/configuration-glossary/configuration-dir/
+  // Directories
   dir: {
     assets: "assets",
-    app: "app",
+    composables: "composables",
     layouts: "layouts",
     middleware: "middleware",
-    store: "store",
+    modules: "modules",
     pages: "pages",
+    plugins: "plugins",
+    public: "../public",
     static: "static"
   },
 
-  // Global CSS: https://go.nuxtjs.dev/config-css
-  css: ["assets/stylesheets/all"],
+  // CSS Files
+  css: ["@/assets/stylesheets/all.scss"],
 
-  // Plugins to run before rendering page: https://go.nuxtjs.dev/config-plugins
-  plugins: [{ src: "plugins/vue-observe-visibility.js" }],
-
-  // Auto import components: https://go.nuxtjs.dev/config-components
+  // Auto-import all components as global
   components: [{ path: "components", pathPrefix: false }],
 
-  // Modules for dev and build (recommended): \
-  //   https://go.nuxtjs.dev/config-modules
-  buildModules: [
-    "@nuxtjs/eslint-module",
-    "@nuxtjs/style-resources",
-    "@nuxtjs/svg"
-  ],
+  // Modules
+  modules: ["@nuxtjs/robots", "@nuxtjs/sitemap", "nuxt-svgo"],
 
-  // Modules: https://go.nuxtjs.dev/config-modules
-  modules: ["@nuxtjs/robots", "@nuxtjs/sitemap"],
-
-  // Router: https://nuxtjs.org/docs/configuration-glossary/configuration-router
-  router: {
-    base: "/",
-    mode: "history",
-    trailingSlash: true,
-    prefetchLinks: false
-  },
-
-  // Loading: \
-  //   https://nuxtjs.org/docs/configuration-glossary/configuration-loading
-  loading: {
-    color: "#ff88a3",
-    failedColor: "#db0031",
-    height: "3px",
-    throttle: 250,
-    duration: 3000,
-    continuous: false
-  },
-
-  // Build Configuration: https://go.nuxtjs.dev/config-build
-  build: {
-    extractCSS: true,
-    publicPath: "/public/",
-
-    splitChunks: {
-      layouts: false,
-      pages: true,
-      commons: true
+  // Preprocessing
+  vite: {
+    css: {
+      preprocessorOptions: {
+        scss: {
+          additionalData: [
+            '@use "@/assets/stylesheets/variables/all.scss" as *;',
+            '@use "@/assets/stylesheets/tools/all.scss" as *;'
+          ].join("\n")
+        }
+      }
     }
   },
 
-  // Public runtime configuration: \
-  //   https://nuxtjs.org/docs/configuration-glossary/\
-  //     configuration-runtime-config
-  publicRuntimeConfig: {
-    // Important: remap config as to strip any private token from there, as \
-    //   eg. in the future there might be some private built-time token shared \
-    //   in this configuration file, which we DO NOT want to leak on the Web.
-    url: CONFIG.url,
-    platforms: CONFIG.platforms,
-    tokens: CONFIG.tokens.public,
-    author: projectPackage.author
+  // Nitro Configuration
+  nitro: {
+    publicDir: "public",
+
+    output: {
+      dir: "build",
+      serverDir: "build/server",
+      publicDir: "build/public"
+    }
   },
 
-  // Generate options: \
-  //   https://nuxtjs.org/docs/configuration-glossary/configuration-generate
-  generate: {
-    dir: "dist",
-    fallback: "404.html"
+  // App Configuration
+  app: {
+    rootId: "__saasmakers",
+    baseURL: "/",
+    buildAssetsDir: "/_saasmakers/",
+
+    // Global page headers
+    head: {
+      meta: [
+        {
+          charset: "utf-8"
+        },
+
+        {
+          name: "viewport",
+          content: "width=device-width, initial-scale=1"
+        },
+
+        {
+          name: "format-detection",
+          content: "telephone=no"
+        },
+
+        {
+          hid: "description",
+          name: "description",
+          content:
+            "Chaque mois, nous parlons d’un sujet SaaS avec un expert du domaine."
+        }
+      ],
+
+      link: [
+        {
+          rel: "shortcut icon",
+          type: "image/x-icon",
+          href: "/favicons/favicon.ico"
+        },
+
+        {
+          rel: "icon",
+          type: "image/png",
+          href: "/favicons/favicon.png"
+        },
+
+        {
+          rel: "icon",
+          type: "image/png",
+          sizes: "512x512",
+          href: "/favicons/favicon-512x512.png"
+        },
+
+        {
+          rel: "icon",
+          type: "image/png",
+          sizes: "256x256",
+          href: "/favicons/favicon-256x256.png"
+        },
+
+        {
+          rel: "icon",
+          type: "image/png",
+          sizes: "192x192",
+          href: "/favicons/favicon-192x192.png"
+        },
+
+        {
+          rel: "icon",
+          type: "image/png",
+          sizes: "144x144",
+          href: "/favicons/favicon-144x144.png"
+        },
+
+        {
+          rel: "icon",
+          type: "image/png",
+          sizes: "128x128",
+          href: "/favicons/favicon-128x128.png"
+        },
+
+        {
+          rel: "icon",
+          type: "image/png",
+          sizes: "32x32",
+          href: "/favicons/favicon-32x32.png"
+        }
+      ]
+    }
   },
 
-  // CLI settings: \
-  //   https://nuxtjs.org/docs/configuration-glossary/configuration-cli
-  cli: {
-    badgeMessages: [
-      `${projectPackage.author.name} © ${new Date().getFullYear()}`
-    ],
-
-    bannerColor: "blueBright"
+  // Runtime Configuration
+  runtimeConfig: {
+    public: {
+      // Important: remap config as to strip any private token from there, as \
+      //   eg. in the future there might be some private built-time token \
+      //   shared in this configuration file, which we DO NOT want to leak on \
+      //   the Web.
+      url: CONFIG.url,
+      platforms: CONFIG.platforms,
+      tokens: CONFIG.tokens.public,
+      author: projectPackage.author
+    }
   },
 
-  // (Module) Robots: https://github.com/nuxt-community/robots-module
-  robots: {
-    UserAgent: "*",
-    Allow: "/",
-    Sitemap: `${CONFIG.url.saasmakers_web}/sitemap.xml`
+  // SVGO Configuration
+  svgo: {
+    autoImportPath: false
   },
 
-  // (Module) Sitemap: https://sitemap.nuxtjs.org/usage/sitemap
+  // Sitemap Configuration
   sitemap: {
     hostname: CONFIG.url.saasmakers_web
   },
 
-  // (Module) Style Resources: \
-  //   https://github.com/nuxt-community/style-resources-module
-  styleResources: {
-    scss: ["assets/stylesheets/variables/*", "assets/stylesheets/tools/*"]
+  // Robots Configuration
+  robots: {
+    rules: {
+      UserAgent: "*",
+      Allow: "/",
+      Sitemap: `${CONFIG.url.saasmakers_web}/sitemap.xml`
+    }
+  },
+
+  // Dev Tools Configuration
+  devtools: {
+    enabled: true
   }
-};
+});
