@@ -14,17 +14,7 @@ page-container
   page-header
 
   page-content
-    nuxt
+    slot
 
   page-footer
 </template>
-
-<!-- **********************************************************************
-     SCRIPT
-     ********************************************************************** -->
-
-<script>
-export default {
-  name: "DefaultLayout"
-};
-</script>
